authConfig rename
